perf

+++ src/inference.py
@@ -7,8 +7,8 @@ from PIL import Image
 import os
 import math
 import time
-# from models import TextEncoder
-from .models import TextEncoder
+from models import TextEncoder
+# from .models import TextEncoder
 from transformers import AutoTokenizer
 try:
     from torchvision.transforms import InterpolationMode
@@ -101,7 +101,7 @@ class Tester():
         batch_input_ids = text_tensor['input_ids']
         batch_attention_mask = text_tensor['attention_mask']
         if isHalf:
-            text_embedding = self.TextEncoder(batch_input_ids.to(self.device), batch_attention_mask.to(self.device)).half()
+            text_embedding = self.TextEncoder(batch_input_ids.to(self.device), batch_attention_mask.to(self.device))
         else:
             text_embedding = self.TextEncoder(batch_input_ids.to(self.device), batch_attention_mask.to(self.device)).float()
         text_embedding = text_embedding / text_embedding.norm(dim=-1, keepdim=True)
@@ -112,7 +112,7 @@ class Tester():
         img = self._preprocess(img)
         with no_grad():
             if isHalf:
-                image_embedding = self.ImageEncoder.encode_image(img.to(self.device)).half() # Output : N x 512
+                image_embedding = self.ImageEncoder.encode_image(img.to(self.device)) # Output : N x 512
             else:
                 image_embedding = self.ImageEncoder.encode_image(img.to(self.device)).float() # Output : N x 512
